Fix dlopen functions
