+++ src/include/extension/extension.h
@@ -8,16 +8,6 @@
 #include "main/db_config.h"
 #include "common/exception/io.h"
 
-#ifdef _WIN32
-
-#include "windows.h"
-#define RTLD_NOW 0
-#define RTLD_LOCAL 0
-
-#else
-#include <dlfcn.h>
-#endif
-
 #define ADD_FUNC(FUNC_STRUCT)                                                                      \
     kuzu::extension::ExtensionUtils::registerFunctionSet(db, std::string(FUNC_STRUCT::name),       \
         FUNC_STRUCT::getFunctionSet())
@@ -137,31 +127,11 @@ struct ExtensionOptions {
 };
 
 #ifdef _WIN32
-std::wstring utf8ToUnicode(const char* input) {
-    uint32_t result;
+std::wstring utf8ToUnicode(const char* input);
 
-    result = MultiByteToWideChar(CP_UTF8, 0, input, -1, nullptr, 0);
-    if (result == 0) {
-        throw common::IOException("Failure in MultiByteToWideChar");
-    }
-    auto buffer = std::make_unique<wchar_t[]>(result);
-    result = MultiByteToWideChar(CP_UTF8, 0, input, -1, buffer.get(), result);
-    if (result == 0) {
-        throw common::IOException("Failure in MultiByteToWideChar");
-    }
-    return std::wstring(buffer.get(), result);
-}
+void* dlopen(const char* file, int /*mode*/);
 
-void* dlopen(const char* file, int /*mode*/) {
-    KU_ASSERT(file);
-    auto fpath = utf8ToUnicode(file);
-    return (void*)LoadLibraryW(fpath.c_str());
-}
-
-void* dlsym(void* handle, const char* name) {
-    KU_ASSERT(handle);
-    return (void*)GetProcAddress((HINSTANCE)handle, name);
-}
+void* dlsym(void* handle, const char* name);
 #endif
 
 } // namespace extension
